refactor: DRY
