+++ src/javascript.rs
@@ -1,6 +1,7 @@
-use npezza93_tree_sitter_tags::{Tag as TSTag, TagsContext, TagsConfiguration};
+use npezza93_tree_sitter_tags::{Tag as TSTag, TagsConfiguration, TagsContext};
 use std::str;
 
+use crate::generate;
 use crate::tag::Tag;
 
 pub fn config() -> TagsConfiguration {
@@ -8,29 +9,26 @@ pub fn config() -> TagsConfiguration {
         tree_sitter_javascript::language(),
         include_str!("../javascript/tags.scm"),
         tree_sitter_javascript::LOCALS_QUERY,
-    ).unwrap()
+    )
+    .unwrap()
 }
 
-pub fn generate_tags<'a>(context: &'a mut TagsContext, config: &'a TagsConfiguration, filename: &'a str, contents: &'a [u8]) -> Vec<Tag> {
-    let tags = context.generate_tags(config, contents, None).unwrap().0;
-
-    tags.flat_map(|tag| {
-        let tag = tag.unwrap();
-        let node_name = config.syntax_type_name(tag.syntax_type_id);
-        let tag_name = &contents[tag.name_range.start..tag.name_range.end];
-        let original_name = str::from_utf8(<&[u8]>::clone(&tag_name)).unwrap_or("");
-
-        vec![create_tag(original_name, node_name, &tag, filename)]
-    }).collect::<Vec<Tag>>()
+pub fn generate_tags<'a>(
+    context: &'a mut TagsContext,
+    config: &'a TagsConfiguration,
+    filename: &'a str,
+    contents: &'a [u8],
+) -> Vec<Tag> {
+    generate::default_generate_tags(create_tag, context, config, filename, contents)
 }
 
 fn create_tag<'a>(name: &'a str, node_name: &'a str, tag: &'a TSTag, filename: &'a str) -> Tag {
     let row = tag.span.start.row;
 
     let kind = match node_name {
-        "method" | "function"  => "f",
+        "method" | "function" => "f",
         "class" => "c",
-        _ => node_name
+        _ => node_name,
     };
 
     Tag::new(name, filename, row + 1, kind)
